Got it working.
I had to install noVNC on the Nginx proxy server. Only the websockify daemon needs to run on the user's instance now. The HTML, JavaScript and other static content are all served directly from the Nginx proxy server. Note that the URL now has 'vnc_auto.html' instead of 'vnc.html'. Something like:

http://kurtz.iplantcollaborative.org/vncws/vnc_auto.html?sig=asdfasdfasdfasdfadfa

I also had to take the 'accept' header out of the browser fingerprint equation because it would vary depending on what was requested.

Also note how I set the cookie after the first authentication. This is then used when authenticating the websocket connection, because the querystring is not passed along. @lenards: I reckon we should just go straight to a cookie and avoid the querystring altogether.