Problem: Proxying websockify fails
Solution: After the recent refactor we did not send back headers to
Nginx any more. Just reinstated the missing headers.

Notes:

We saw this in the Nginx log files:

```
no host in upstream ":4200" while sending to client
```

This was a clue in the `novnc_auth.log`:

```
Sending back headers: {}
```

The crucial missing header was the `X-Target-VM-IP` header which
contained the upstream VM IP.